Replaced the `Document` model's `id` attribute with `docid` as I feel it more accurately conveys what it is (`key` was also a contender).
In addition to bringing more clarity it frees up an `id` member for
implementations to make use of. Consider the following scenario:

    Index prefix: `card:`
    Document key: `card:5`

The *actual* id of the document is likely everything to the right of the
colon. Using `id` for the entire document key made it so that
implementers would have to use some ambiguous member name to return the
extracted `5` from the above key.